requirements.txt
APScheduler==3.8.1
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cloudinary==1.28.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
numpy==1.22.2
pandas==1.4.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==5.3.0
wincertstore==0.2
